add more tests to message-relayer-fast
