+++ packages/omgx/message-relayer-fast/.env.example
@@ -10,3 +10,5 @@ TARGET_GAS_LIMIT=9000000000
 CHAIN_ID=28
 L1_WALLET_KEY=0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80
 L1_TARGET=0x0
+URL=http://localhost:8080/addresses.json
+OMGX_URL=http://localhost:8078/addresses.json
